requirements.txt
spotipy==2.22.1
gensim==4.3.1
nltk==3.8.1
pandas==2.0.3
scikit-learn==1.3.0
textblob==0.17.1
python-dotenv==1.0.0